requirements.txt
altair==5.2.0
blinker==1.8.2
boto3
botocore
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
distro==1.9.0
fastapi==0.108.0
ffmpy==0.3.1
Flask
fsspec==2023.12.2
gradio
gradio_client
gunicorn
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.10
importlib-resources==6.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
langchain-community
langchain-openai
langchain-together
MarkupSafe
marshmallow==3.20.1
openai
packaging==23.2
pydub==0.25.1
python-dotenv
python-dateutil==2.9.0.post0
python-multipart==0.0.6
requests==2.32.3
rich==13.7.0
s3transfer==0.10.3
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sseclient-py==1.8.0
starlette==0.32.0.post1
tabulate==0.9.0
tiktoken==0.6.0
together==0.2.11
tomlkit==0.12.0
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.25.0
websockets==11.0.3
Werkzeug==3.0.4
